+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 0.0.2
+* added customer error message that appears upon failed submit (for form and submit widget)
+* Added loading param for form
 
-* TODO: Describe initial release.
+
+## 0.0.1
+* init